desloppify: move loss implementation into objectives
